requirements.txt
fastapi==0.115.11
passlib==1.7.4
pydantic==2.10.6
PyJWT==2.10.1
sqlmodel==0.0.24
starlette==0.46.1
fabric~=3.2.2
colorama~=0.4.6
python-dotenv~=1.0.1
httpx~=0.28.1
sqlalchemy~=2.0.39
uvicorn==0.34.0
python-multipart==0.0.20

